# server/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Videofly API",
    "version": "0.0.1",
    "description": "Videofly backend API for authentication, user management, video uploads, and video playback."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment origin"
    }
  ],
...
  "paths": {
...
    "/api/v1/upload/init": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Initialize a multipart upload",
        "description": "Uploader and admin only.",
        "operationId": "initUpload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSession"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or unsupported video type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Quota exceeded or filename conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/part": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Mark a multipart upload as in progress",
        "description": "Uploader and admin only.",
        "operationId": "prepareUploadPart",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadPartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload session state updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPartStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Upload session cannot accept parts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/part/upload": {
      "put": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload a single binary part",
        "description": "Uploader and admin only. Send raw bytes in the request body.",
        "operationId": "uploadPartBinary",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "x-part-sha256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lowercase SHA-256 checksum of the raw chunk."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload session updated with uploaded part",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid chunk or checksum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Upload session metadata missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OSS upload failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/complete": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Complete a multipart upload",
        "description": "Uploader and admin only.",
        "operationId": "completeUpload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedVideo"
                }
              }
            }
          },
          "400": {
            "description": "Upload is incomplete or cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Upload session expired in OSS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/cancel": {
      "delete": {
        "tags": [
          "Upload"
        ],
        "summary": "Cancel a multipart upload",
        "description": "Uploader and admin only.",
        "operationId": "cancelUpload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload canceled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanceledUpload"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/history": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "List recent upload sessions",
        "description": "Uploader and admin only.",
        "operationId": "listUploadHistory",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Upload history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/status/{uploadId}": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "Get upload session status",
        "description": "Uploader and admin only.",
        "operationId": "getUploadStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Upload session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSession"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "PlaybackToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "schemas": {
...
      "InitUploadRequest": {
        "type": "object",
        "required": [
          "title",
          "filename",
          "mimeType",
          "fileSizeBytes"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "filename": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "video/mp4",
              "video/quicktime",
              "video/x-msvideo",
              "video/x-matroska"
            ]
          },
          "fileSizeBytes": {
            "oneOf": [
              {
                "type": "string",
                "pattern": "^[0-9]+$"
              },
              {
                "type": "integer",
                "minimum": 1
              }
            ]
          }
        }
      },
...
      "CompleteUploadRequest": {
        "type": "object",
        "required": [
          "uploadId"
        ],
        "properties": {
          "uploadId": {
            "type": "string"
          }
        }
      },
...
      "UploadSession": {
        "type": "object",
        "required": [
          "uploadId",
          "status",
          "partSizeBytes",
          "fileSizeBytes",
          "uploadedParts",
          "uploadedPartDetails",
          "totalParts"
        ],
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UploadSessionStatus"
          },
          "partSizeBytes": {
            "type": "integer",
            "minimum": 1
          },
          "fileSizeBytes": {
            "type": "string",
            "example": "10485760"
          },
          "uploadedParts": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1
            }
          },
          "uploadedPartDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPartDetail"
            }
          },
          "totalParts": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}